+++ RgbColor.cs
@@ -4,11 +4,11 @@ namespace KernelConvolution
 {
     public class RgbColor
     {
-        private int r { get; set; }
-        private int g { get; set; }
-        private int b { get; set; }
+        private double r { get; set; }
+        private double g { get; set; }
+        private double b { get; set; }
 
-        public RgbColor(int r, int g, int b)
+        public RgbColor(double r, double g, double b)
         {
             this.r = r;
             this.g = g;
@@ -23,15 +23,15 @@ namespace KernelConvolution
         public static Color ConvertoToColor(RgbColor c)
         {
             Color result = Color.FromArgb(
-                c.r < 0 ? 0 : c.r > 255 ? 255 : c.r,
-                c.g < 0 ? 0 : c.g > 255 ? 255 : c.g,
-                c.b < 0 ? 0 : c.b > 255 ? 255 : c.b
+                c.r < 0 ? 0 : c.r > 255 ? 255 : (int)c.r,
+                c.g < 0 ? 0 : c.g > 255 ? 255 : (int)c.g,
+                c.b < 0 ? 0 : c.b > 255 ? 255 : (int)c.b
             );
 
             return result;
         }       
         
-        public static RgbColor MultColorByNum(Color c, int num)
+        public static RgbColor MultColorByNum(Color c, double num)
         {
             return new RgbColor(c.R * num, c.G * num, c.B * num);
         }
